requirements.txt
streamlit>=1.35.0
streamlit-authenticator>=0.3.2
gspread>=6.0.0
google-auth>=2.28.0
pandas>=2.0.0
plotly>=5.18.0
pyyaml>=6.0
bcrypt>=4.1.0
openpyxl>=3.1.0
